Add help text and a dry run flag to gather export

+++ builders/flight-gather/libexec/export.sh
@@ -25,6 +25,42 @@
 # https://github.com/alces-flight/alces-flight-omnibus-builder
 #===============================================================================
 
+usage() {
+    local prog
+    prog="flight gather export"
+    cat <<USAGE
+Usage: ${prog} [--help] [--dry-run] ASSET_NAME...
+Export inventory data to Flight Center
+  --help    Display this help text
+  --dry-run Output the asset sheets to a directory, instead of Flight Center
+USAGE
+}
+
+# Parse the help and dry-run flags
+PARAMS=""
+DRY_RUN=""
+while (( "$#" )); do
+  case "$1" in
+    -h|--help)
+      usage
+      exit
+      ;;
+    --dry-run)
+      DRY_RUN="true"
+      shift
+      ;;
+    -*|--*=) # unsupported flags
+      echo "Error: Unsupported flag $1" >&2
+      exit 1
+      ;;
+    *) # preserve positional arguments
+      PARAMS="$PARAMS $1"
+      shift
+      ;;
+  esac
+done
+eval set -- "$PARAMS"
+
 # Ensures flight-asset has been configured
 flight asset list 2>/dev/null >&2
 if [ $? -ne 0 ]; then
@@ -52,51 +88,61 @@ for asset in "$@"; do
   fi
 done
 
-# Uploads the info
-for path in *; do
-  asset=$(basename $path)
 
-  # Attempts an update
-  flight asset update $asset --info @$path 2>/dev/null >&2
-  case $? in
-  0)
-    echo "Exported (update): $asset"
-    ;;
-  21)
-    # Attempts a create
-    flight asset create $asset --info @$path 2>/dev/null >&2
-    if [ $? -eq 0 ]; then
-      echo "Exported (create): $asset"
-    else
+# For "Dry Runs" output the directory where the render files are stored
+if [ "$DRY_RUN" ]; then
+  cat <<INFO
+The rendered asset sheets can be found in:
+$(pwd)
+INFO
+
+# Uploads the asset info
+else
+  for path in *; do
+    asset=$(basename $path)
+
+    # Attempts an update
+    flight asset update $asset --info @$path 2>/dev/null >&2
+    case $? in
+    0)
+      echo "Exported (update): $asset"
+      ;;
+    21)
+      # Attempts a create
+      flight asset create $asset --info @$path 2>/dev/null >&2
+      if [ $? -eq 0 ]; then
+        echo "Exported (create): $asset"
+      else
+        echo "Failed to export: $asset"
+        exit_code=1
+      fi
+      ;;
+    *)
       echo "Failed to export: $asset"
       exit_code=1
-    fi
-    ;;
-  *)
-    echo "Failed to export: $asset"
-    exit_code=1
-    ;;
-  esac
-done
+      ;;
+    esac
+  done
 
-# Remove the temporary directory
-popd >/dev/null 2>&1
-rm -rf $local_dir
+  # Remove the temporary directory
+  popd >/dev/null 2>&1
+  rm -rf $local_dir
 
-# Determines assets with missing groups
-sorted_assets=$(echo "$@" | xargs -n 1 | sort)
-sorted_missing=$(flight asset list-assets --group '' | cut -f1 | xargs -n1 | sort)
-missing=$(comm -12 <(echo $sorted_assets | xargs -n1) <(echo $sorted_missing | xargs -n1))
+  # Determines assets with missing groups
+  sorted_assets=$(echo "$@" | xargs -n 1 | sort)
+  sorted_missing=$(flight asset list-assets --group '' | cut -f1 | xargs -n1 | sort)
+  missing=$(comm -12 <(echo $sorted_assets | xargs -n1) <(echo $sorted_missing | xargs -n1))
 
-if [ -n "$missing" ]; then
-  cat <<HERE >&2
+  if [ -n "$missing" ]; then
+    cat <<HERE >&2
 
 The following assets have not been assigned to a group:
 $(echo $missing | xargs)
 
 You may add them to a group with:
-flight asset move ASSET GROUP
+$flight_ROOT/bin/flight asset move ASSET GROUP
 HERE
+  fi
 fi
 
 exit $exit_code
